Update base version
Completed the Assignment's iteration 3 by adding an image, different heading tags, and lists to the recipe file (../recipes/carrotcakepancakes.html). Configured Prettier in VSC to hard wrap column width to 80 which is the conventional characters limit in a line of code. Note: If it does not automatically hard wrap, Press ALT+SHIFT+F